Move flagChanged to Socket from OutputSocket
Call flagChanged in ProjectTest
